+++ R/db_info.R
@@ -4,7 +4,7 @@
 #' @param ... force all arguments to be by name.
 #' @param connection connection handle to database or Spark.
 #' @param is_dbi if TRUE the database connection can be used with DBI.
-#' @param indentifier_quote_char character, quote to put around identifiers.
+#' @param identifier_quote_char character, quote to put around identifiers.
 #' @param string_quote_char character, quote to put around strings.
 #' @param overrides named list of functions to place in info.
 #' @param note character note to add to display form.
@@ -16,7 +16,7 @@
 rquery_db_info <- function(...,
                            connection = NULL,
                            is_dbi = FALSE,
-                           indentifier_quote_char = NULL,
+                           identifier_quote_char = NULL,
                            string_quote_char = NULL,
                            overrides = NULL,
                            note = "",
@@ -29,14 +29,20 @@ rquery_db_info <- function(...,
   r <- list(
     connection = connection,
     is_dbi = is_dbi,
-    indentifier_quote_char = indentifier_quote_char,
+    identifier_quote_char = identifier_quote_char,
     string_quote_char = string_quote_char,
     note = note,
     connection_options = connection_options,
     dbqi = function(id) {
-      paste0(indentifier_quote_char,
+      paste0(identifier_quote_char,
              id,
-             indentifier_quote_char)
+             identifier_quote_char)
+    },
+    dbqt = function(id, qualifiers = NULL) {
+      paste(paste0(identifier_quote_char,
+                   c(qualifiers, id),
+                   identifier_quote_char),
+            collapse = ".")
     },
     dbqs = function(s) {
       paste0(string_quote_char,
@@ -56,7 +62,17 @@ rquery_db_info <- function(...,
       stop("rquery::rquery_db_info is_dbi=TRUE requeries DBI package")
     }
     r$quote_identifier <- function(x, id) {
-      DBI::dbQuoteIdentifier(r$connection, as.character(id))
+      as.character(DBI::dbQuoteIdentifier(r$connection, as.character(id)))
+    }
+    r$quote_table_name <- function(x, id, ..., qualifiers) {
+      wrapr::stop_if_dot_args(substitute(list(...)),
+                              "r$quote_table_name")
+      if("schema" %in% names(qualifiers)) {
+        dbi_id <- DBI::Id(schema = qualifiers$schema, table = as.character(id))
+      } else {
+        dbi_id <- as.character(id) # sparklyr ‘0.8.4’ does not implement DBI::dbQuoteIdentifier for DBI::Id
+      }
+      as.character(DBI::dbQuoteIdentifier(r$connection, dbi_id))
     }
     r$quote_string <- function(x, s) {
       DBI::dbQuoteString(r$connection, as.character(s))
@@ -65,7 +81,7 @@ rquery_db_info <- function(...,
       if(is.character(o) || is.factor(o)) {
         return(DBI::dbQuoteString(r$connection, as.character(o)))
       }
-      DBI::dbQuoteLiteral(r$connection, o)
+      as.character(DBI::dbQuoteLiteral(r$connection, o))
     }
   }
   for(ni in names(overrides)) {
@@ -78,7 +94,7 @@ rquery_db_info <- function(...,
 #' @export
 format.rquery_db_info <- function(x, ...) {
   wrapr::stop_if_dot_args(substitute(list(...)), "rquery::format.rquery_db_info")
-  paste0("rquery_db_info(is_dbi=", x$is_dbi, ", ", x$note, ", ", format(x$connection) , ")")
+  paste0("rquery_db_info(",rq_connection_name(x), ", is_dbi=", x$is_dbi, ", note=\"", x$note, "\")")
 }
 
 #' @export
@@ -88,7 +104,7 @@ print.rquery_db_info <- function(x, ...) {
 }
 
 
-rquery_default_db_info <- rquery_db_info(indentifier_quote_char = '"',
+rquery_default_db_info <- rquery_db_info(identifier_quote_char = '"',
                                          string_quote_char = "'",
                                          is_dbi = FALSE)
 
@@ -100,7 +116,7 @@ rquery_default_db_info <- rquery_db_info(indentifier_quote_char = '"',
 #'
 #' @export
 #'
-quote_identifier <- function (x, id) {
+quote_identifier <- function(x, id) {
   if("rquery_db_info" %in% class(x)) {
     f <- x$quote_identifier
     if(!is.null(f)) {
@@ -114,6 +130,40 @@ quote_identifier <- function (x, id) {
   rquery_default_db_info$dbqi(id)
 }
 
+#' Quote a table name.
+#'
+#' @param x database handle or rquery_db_info object.
+#' @param id character to quote
+#' @param ... not used, force later arguments to bind by name.
+#' @param qualifiers named ordered vector of strings carrying additional db hierarchy terms, such as schema.
+#' @return quoted identifier
+#'
+#' @export
+#'
+quote_table_name <- function(x, id,
+                             ...,
+                             qualifiers) {
+  wrapr::stop_if_dot_args(substitute(list(...)),
+                          "rquery::quote_table_name")
+  if("rquery_db_info" %in% class(x)) {
+    f <- x$quote_table_name
+    if(!is.null(f)) {
+      return(f(x, id, qualifiers = qualifiers))
+    }
+    return(x$dbqt(id, qualifiers = qualifiers))
+  }
+  if(requireNamespace("DBI", quietly = TRUE)) {
+    if("schema" %in% names(qualifiers)) {
+      dbi_id <- DBI::Id(schema = qualifiers[["schema"]], table = as.character(id))
+    } else {
+      dbi_id <- as.character(id) # sparklyr ‘0.8.4’ does not implement DBI::dbQuoteIdentifier for DBI::Id
+    }
+    return(as.character(DBI::dbQuoteIdentifier(x, dbi_id)))
+  }
+  rquery_default_db_info$dbqi(id)
+}
+
+
 #' Quote a string
 #'
 #' @param x database handle or rquery_db_info object.
@@ -122,7 +172,7 @@ quote_identifier <- function (x, id) {
 #'
 #' @export
 #'
-quote_string <- function (x, s) {
+quote_string <- function(x, s) {
   s <- as.character(s)
   if("rquery_db_info" %in% class(x)) {
     f <- x$quote_string
@@ -145,7 +195,7 @@ quote_string <- function (x, s) {
 #'
 #' @export
 #'
-quote_literal <- function (x, o) {
+quote_literal <- function(x, o) {
   if(is.character(o) || is.factor(o)) {
     return(quote_string(x, as.character(o)))
   }